+++ packages/plugin-analytic-fb-pixel/src/main.ts
@@ -1,27 +1,9 @@
 import { pluginLoader } from '@ima/core';
 
-import FacebookPixelAnalytic from './FacebookPixelAnalytic';
-
-export interface PluginAnalyticFBPixelSettings {
-  fbPixel: {
-    id: string | null;
-  };
-}
-
-declare module '@ima/core' {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface PluginAnalyticSettings extends PluginAnalyticFBPixelSettings {}
-
-  interface PluginSettings {
-    analytic: PluginAnalyticSettings;
-  }
-
-  interface Settings {
-    plugin: PluginSettings;
-  }
-}
-
-const defaultDependencies = FacebookPixelAnalytic.$dependencies;
+import {
+  FacebookPixelAnalytic,
+  type AnalyticFBPixelSettings,
+} from './FacebookPixelAnalytic';
 
 pluginLoader.register('@ima/plugin-analytic-google', () => ({
   initSettings: () => ({
@@ -37,4 +19,5 @@ pluginLoader.register('@ima/plugin-analytic-google', () => ({
   }),
 }));
 
-export { FacebookPixelAnalytic, defaultDependencies };
+export type { PluginAnalyticFBPixelSettings } from './types';
+export { FacebookPixelAnalytic, type AnalyticFBPixelSettings };