+++ packages/plugin-analytic-google/src/__tests__/GoogleAnalytics4Spec.js
@@ -2,7 +2,7 @@ import { Dispatcher, Window } from '@ima/core';
 import { ScriptLoaderPlugin } from '@ima/plugin-script-loader';
 import { toMockedInstance } from 'to-mock';
 
-import GoogleAnalytics4 from '../GoogleAnalytics4';
+import { GoogleAnalytics4 } from '../GoogleAnalytics4';
 
 describe('GoogleAnalytics4', () => {
   const settings = {
@@ -42,10 +42,10 @@ describe('GoogleAnalytics4', () => {
     });
 
     googleAnalytics4 = new GoogleAnalytics4(
+      settings,
       scriptLoader,
       window,
-      dispatcher,
-      settings
+      dispatcher
     );
   });