+++ packages/plugin-analytic-google/src/main.ts
@@ -1,33 +1,9 @@
-/* eslint-disable @typescript-eslint/no-empty-interface */
 import { pluginLoader } from '@ima/core';
 
-import GoogleAnalytics4 from './GoogleAnalytics4.js';
-
-const googleAnalytics4DefaultDependencies = GoogleAnalytics4.$dependencies;
-
-export interface PluginAnalyticGoogleSettings {
-  google4: {
-    consentSettings?: {
-      ad_storage?: 'denied' | 'granted';
-      analytics_storage?: 'denied' | 'granted';
-      personalization_storage?: 'denied' | 'granted';
-    };
-    service: string;
-    waitForUpdateTimeout?: number;
-  };
-}
-
-declare module '@ima/core' {
-  interface PluginAnalyticSettings extends PluginAnalyticGoogleSettings {}
-
-  interface PluginSettings {
-    analytic: PluginAnalyticSettings;
-  }
-
-  interface Settings {
-    plugin: PluginSettings;
-  }
-}
+import {
+  GoogleAnalytics4,
+  type AnalyticGoogleSettings,
+} from './GoogleAnalytics4';
 
 pluginLoader.register('@ima/plugin-analytic-google', () => ({
   initSettings: () => ({
@@ -52,4 +28,5 @@ pluginLoader.register('@ima/plugin-analytic-google', () => ({
   }),
 }));
 
-export { GoogleAnalytics4, googleAnalytics4DefaultDependencies };
+export type { PluginAnalyticGoogleSettings } from './types';
+export { GoogleAnalytics4, type AnalyticGoogleSettings };