+++ packages/plugin-analytic/src/main.ts
@@ -1,25 +1,9 @@
-import AbstractAnalytic from './AbstractAnalytic';
+import './types';
+import {
+  AbstractAnalytic,
+  type InitConfig as AbstractAnalyticInitConfig,
+} from './AbstractAnalytic';
 import { Events } from './Events';
 
-const defaultDependencies = AbstractAnalytic.$dependencies;
-
-declare module '@ima/core' {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface PluginAnalyticSettings {}
-
-  interface PluginSettings {
-    analytic: PluginAnalyticSettings;
-  }
-
-  interface Settings {
-    plugin: PluginSettings;
-  }
-
-  interface DispatcherEventsMap {
-    [Events.LOADED]: {
-      type: string;
-    };
-  }
-}
-
-export { Events, AbstractAnalytic, defaultDependencies };
+export type { AbstractAnalyticInitConfig };
+export { Events, AbstractAnalytic };