+++ packages/plugin-atoms/src/types.ts
@@ -1,0 +1,9 @@
+import type UIComponentHelper from './UIComponentHelper';
+
+declare module '@ima/core' {
+  interface Utils {
+    $UIComponentHelper: UIComponentHelper;
+  }
+}
+
+export {};