+++ packages/plugin-script-loader/src/main.ts
@@ -1,15 +1,35 @@
+import './types';
+import { pluginLoader, ComponentUtils } from '@ima/core';
+
 import { Events } from './Events';
-import ScriptLoaderPlugin from './ScriptLoaderPlugin';
+import { ScriptLoader } from './ScriptLoader';
 
-declare module '@ima/core' {
-  interface DispatcherEventsMap {
-    [Events.LOADED]: {
-      url: string;
-      error?: Error;
-    };
-  }
-}
+const defaultDependencies = ScriptLoader.$dependencies;
 
-const defaultDependencies = ScriptLoaderPlugin.$dependencies;
+pluginLoader.register('@ima/plugin-script-loader', () => ({
+  initBind: (ns, oc) => {
+    oc.get(ComponentUtils).register(
+      'ScriptLoader',
+      ScriptLoader,
+      '@ima/plugin-script-loader'
+    );
 
-export { Events, ScriptLoaderPlugin, defaultDependencies };
+    /**
+     * Deprecated, use 'ScriptLoader' instead, this is here
+     * for backwards compatibility.
+     */
+    oc.get(ComponentUtils).register(
+      'ScriptLoaderPlugin',
+      ScriptLoader,
+      '@ima/plugin-script-loader'
+    );
+  },
+}));
+
+export {
+  Events,
+  ScriptLoader,
+  // @deprecated, don't use this alias
+  ScriptLoader as ScriptLoaderPlugin,
+  defaultDependencies,
+};