+++ packages/cli-plugin-scramble-css/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.2.7
+
+### Patch Changes
+
+- 8f5d01a: Fixed build issues with client server config
+
+## 1.2.6
+
+### Patch Changes
+
+- e2eeb2b: Turbo, eslint update, types exports, new tsconfig
+
 ## 1.2.5
 
 ### Patch Changes