+++ packages/plugin-http-client/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ima/plugin-http-client
 
+## 2.1.3
+
+### Patch Changes
+
+- 8f5d01a: Fixed build issues with client server config
+
+## 2.1.2
+
+### Patch Changes
+
+- e2eeb2b: Turbo, eslint update, types exports, new tsconfig
+
+## 2.1.1
+
+### Patch Changes
+
+- 6d92917: Code refactorings, moved type extensions to separate file to cleanup main.ts
+
 ## 2.1.0
 
 ### Minor Changes