+++ packages/plugin-resource-loader/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 3.0.4
+
+### Patch Changes
+
+- 8f5d01a: Fixed build issues with client server config
+
+## 3.0.3
+
+### Patch Changes
+
+- 038a8ea: Typescriptization (from .js to .ts)
+
 ## 3.0.2
 
 ### Patch Changes